pyproject.toml
[project]
name = "finance-bench"
version = "0.1.0"
description = "FinanceBench document processing system for downloading, converting, and segmenting financial documents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mistralai>=1.0.0",
    "openai>=1.0.0",
    "instructor>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.25.0",
    "tqdm>=4.60.0",
    "pathlib>=1.0.0",
]

[project.scripts]
download-pdfs = "run_download:main"
run-ocr = "run_ocr:main"
run-segmentation = "run_segmentation:main"
run-pipeline = "run_pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
